server/requirements.txt
fastapi
uvicorn
websockets
numpy
opencv-python
mediapipe
google-generativeai
gTTS
python-dotenv

# Machine Learning
tensorflow>=2.10.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Dataset handling
datasets>=2.14.0
huggingface-hub>=0.16.0

# For TensorFlow.js conversion (optional)
tensorflowjs>=4.10.0